Fourier transform of basis elements

LinearResponse.Wintegrand — Function

Wintegrand(w,a,e,L,Ω1,Ω2,ψ,dψ,d2ψ,basis,params)

Integrand computation/update for FT of basis elements

LinearResponse.WBasisFT — Function

WBasisFT(a,e,Ω1,Ω2,n1,n2,)

Fourier Transform of basis elements using RK4 scheme result stored in place

with basisFT struct

without Ω1, Ω2

Integral over the resonance line

LinearResponse.MakeGu — Function

MakeGu(ndFdJ,n1,n2,Wdata,tabu[,params])

function to compute G(u)

Decomposition

LinearResponse.RunAXi — Function

RunAXi(FHT,params)

function to make the decomposition coefficients "a" of the response matrix M

these values do not depend on the frequency being evaluated: which makes them good to table

is this struggling from having to pass around a gigantic array? what if we did more splitting?

Response matrix computations

LinearResponse.tabM! — Function

tabM!(ω,tabM,tabaMcoef,tabωminωmax,FHT,params)

computes the response matrix M(ω) for a given COMPLEX frequency ω in physical units, i.e. not (yet) rescaled by 1/Ω0.

@IMPROVE: The shape of the array could maybe be improved

See LinearTheory.jl for a similar version